#18 database level beta

+++ translator-cli-meta/translator-cli-words-normalize/src/main/java/com/github/bogdanovmn/translator/cli/wordsnormilize/WordsNormalizeService.java
@@ -5,6 +5,8 @@ import com.github.bogdanovmn.translator.web.orm.Word;
 import com.github.bogdanovmn.translator.web.orm.WordRepository;
 import com.github.bogdanovmn.translator.web.orm.WordSource;
 import com.github.bogdanovmn.translator.web.orm.WordSourceRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,11 +18,13 @@ import java.util.stream.Collectors;
 
 @Service
 class WordsNormalizeService {
+	private static final Logger LOG = LoggerFactory.getLogger(WordsNormalizeService.class);
+
 	private final WordRepository wordRepository;
 	private final WordSourceRepository wordSourceRepository;
 
 	@Autowired
-	public WordsNormalizeService(WordRepository wordRepository, WordSourceRepository wordSourceRepository) {
+	WordsNormalizeService(WordRepository wordRepository, WordSourceRepository wordSourceRepository) {
 		this.wordRepository = wordRepository;
 		this.wordSourceRepository = wordSourceRepository;
 	}
@@ -38,33 +42,44 @@ class WordsNormalizeService {
 
 	@Transactional
 	synchronized void normalize() {
+		LOG.info("Start normalize process");
+
 		Map<String, Word> wordsMap = this.wordRepository.getAllByBlackListFalse().stream()
 			.collect(Collectors.toMap(
 				Word::getName, x -> x
 			));
 
+		LOG.info("Total words before: {}", wordsMap.keySet().size());
+
 		NormalizedWords normalizedWords = new NormalizedWords(
 			wordsMap.keySet()
 		);
 
+		LOG.info("Normalized words: {}", normalizedWords.get().size());
+
 		for (String normal : normalizedWords.get()) {
 			Set<String> forms = normalizedWords.getFormsForWord(normal);
-			Word normalWord = wordsMap.get(normal);
+			if (!forms.isEmpty()) {
+				LOG.info("Word '{}' has forms: {}", normal, forms);
 
-			for (String form : forms) {
-				Word formWord = wordsMap.get(form);
-				Set<WordSource> formSources = formWord.getWordSources();
+				Word normalWord = wordsMap.get(normal);
 
-				normalWord.incFrequence(formWord.getFrequence());
-				mergeSources(normalWord, formSources);
-				wordRepository.delete(formWord);
+				for (String form : forms) {
+					Word formWord = wordsMap.get(form);
+					Set<WordSource> formSources = wordSourceRepository.findAllByWord(formWord);
+
+					normalWord.incFrequence(formWord.getFrequence());
+					mergeSources(normalWord, formSources);
+					wordRepository.delete(formWord);
+				}
+				wordRepository.save(normalWord);
 			}
-			wordRepository.save(normalWord);
 		}
+		LOG.info("Finish normalize process");
 	}
 
 	private void mergeSources(Word word, Set<WordSource> formSources) {
-		Set<WordSource> wordSources = word.getWordSources();
+		Set<WordSource> wordSources = wordSourceRepository.findAllByWord(word);
 
 		Map<Integer, WordSource> wordSourceMap = getSourceMap(wordSources);
 		Map<Integer, WordSource> formSourceMap = getSourceMap(formSources);
@@ -73,11 +88,13 @@ class WordsNormalizeService {
 			(id, formSource) -> {
 				// Слово и его форма из одного источника
 				if (wordSourceMap.containsKey(id)) {
+					LOG.info("Same source '{}', increment source word counter", id);
 					wordSourceMap.get(id).incCount(
 						formSource.getCount()
 					);
 				}
 				else {
+					LOG.debug("New source '{}', add it for word", id);
 					wordSources.add(
 						new WordSource()
 							.setWord(word)
